requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
Django==3.2.12
django-bootstrap5==21.2
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-slugger==1.1.2
gunicorn==20.1.0
importlib-metadata==2.1.3
Pillow==8.4.0
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.6.0
